requirements.txt
Django==5.2
psycopg2-binary==2.9.10
gunicorn==23.0.0
whitenoise==6.9.0
django-pwa==2.0.1
pillow==11.1.0
python-decouple==3.8
